Removed handling of legacy names from server code.
I removed the code from the JSON client code as well as the Sample class in the
core of the remote client.  This way, each institute must deal with legacy
sample names in their remote client code.  jb_institute still contains the
Legacy topic and the user nobody which own the LGCY initials.  This is the
approach that we recommend in JuliaBase.