Removed all the heuristics code around sample and substrate in the remote client.
This affects the creation of a new sample.  Formerly (this was IEK-5 code),
the client heavily guessed which is the real sampole name, the sample
topic, the current location, etc.  Furthermore, the substrate was derived
from possibly misspelled substrate remarks in data files.  All this was removed
to make jb_remote_institute.py didactically clearer and more useful.