Implemented new policy on how to use ugettext and ugettext_lazy.
Now, at the beginning of a module, _ is ugettext_lazy, and at the end,
ugettext.  This way, it is ugettext_lazy is long as no language has been set
(during initialization), and ugettext in functions and methods when called
after the actual request, i.e. when a lenguage is set.