Implemented first changes towards a splitting of DepositionView into mixins.
It is not working yet.  In particular, the new class-based views for
depositions are simply copies of the mixins, which means that the strings
(e.g. error messages) speak of "steps" instead of "layers".  Moreover, the
ability to copy initial data from an existing process should be extracted into
a mixin of its own.  And finally, the views haven't been updated yet.