Switched off flags for unauthenticated users.
Firstly, it does not make sense for unauthenticated users: You need to be
logged-in for having user details in which the language can be saved.  (This
may be a poor constraint but there we go.)  And secondly, at least one
bot (SEMrushBot) gets confused by this and follows the language URL/login URL,
getting arbitrarily long URLs.  This caused a DoS in one case.