Add the “sites” framework
It should be non-invasive for existing installations.  The usual migrations add
the tables necessary for the sites framework, and add a site with site ID 1 and
the domain and name “example.com”.

I need the sites framework for the knowledge graph extension because I need the
domain name of the current site.  The are various ways to get this domain, but
the sites framework seems to be the cleanest one.  Besides, it is a very small
change overall.