Use “title” attribute for process entry on sample data sheet
Previously, the “Result” process was special-cased.  Now, I use duck typing: If
the process has got a “title” field (which uses to be the case only for the
Result process), it is used for the heading of the process entry on the sample
data sheet.

In the future, at least the “external data” process will also have a title
field.  So, I am more flexible with the new code.