Add ExternalData model
Note that this model is not managed by Django (“managed” is false).  I only
need the model functionality but (at least, for the time being) don’t want to
have it written to the database.  Instead, its data is fetched from the remote
site every time the data sheet of the sample is compiled.

Of course, this code is preliminary.  It only works if the remote-sample data
is local because it is fetched via the remote client.  This is because I need
to evade login-based authentication.

The next big thing is to display sub-experiments (layers).